rename Python module to attribute_ruler
